+++ tests/static/test_cfg_clients.py
@@ -58,6 +58,42 @@ class TestConfigParserClients(TestConfigParserBase):
         self.cfg[1]["ip"] = self.cfg[0]["ip"]
         self._contains_err("Duplicate IP address found: 192.0.2.1.")
 
+    def test_next_hop_policy_no_authorized_addresses_with_list(self):
+        """{}: next_hop.policy != authorized_addresses, list given"""
+        self.cfg[0]["cfg"]["filtering"]["next_hop"]["policy"] = "strict"
+        self.cfg[0]["cfg"]["filtering"]["next_hop"]["authorized_addresses_list"] = ["192.168.0.1"]
+        self._contains_err("AS3333 192.0.2.11 is not 'authorized_addresses' but the 'authorized_addresses_list' option is set")
+
+    def test_next_hop_policy_authorized_addresses_empty(self):
+        """{}: next_hop.policy authorized_addresses, empty list"""
+        self.cfg[0]["cfg"]["filtering"]["next_hop"]["policy"] = "authorized_addresses"
+        self._contains_err("set to 'authorized_addresses' but the list of authorized IP addresses ('authorized_addresses_list') is empty")
+
+    def test_next_hop_policy_authorized_addresses_invalid(self):
+        """{}: next_hop.policy authorized_addresses, invalid list"""
+        self.cfg[0]["cfg"]["filtering"]["next_hop"]["policy"] = "authorized_addresses"
+        self.cfg[0]["cfg"]["filtering"]["next_hop"]["authorized_addresses_list"] = "test"
+        self._contains_err("Error parsing 'authorized_addresses_list' at 'clients.cfg.filtering.next_hop' level - Invalid format: must be a list.")
+
+    def test_next_hop_policy_authorized_addresses_invalid_address(self):
+        """{}: next_hop.policy authorized_addresses, invalid IP addr"""
+        self.cfg[0]["cfg"]["filtering"]["next_hop"]["policy"] = "authorized_addresses"
+        self.cfg[0]["cfg"]["filtering"]["next_hop"]["authorized_addresses_list"] = ["test"]
+        self._contains_err("Error parsing 'authorized_addresses_list' at 'clients.cfg.filtering.next_hop' level - Invalid IP address: test.")
+
+    def test_next_hop_policy_both(self):
+        """{}: next_hop.policy and next_hop_policy"""
+        self.cfg[0]["cfg"]["filtering"]["next_hop_policy"] = "strict"
+        self.cfg[0]["cfg"]["filtering"]["next_hop"]["policy"] = "authorized_addresses"
+        self.cfg[0]["cfg"]["filtering"]["next_hop"]["authorized_addresses_list"] = ["test"]
+        self._contains_err("Can't use the new 'next_hop' and the old 'next_hop_policy' statements simultaneously")
+
+    def test_next_hop_policy_authorized_addresses_ok(self):
+        """{}: next_hop.policy authorized_addresses, ok"""
+        self.cfg[0]["cfg"]["filtering"]["next_hop"]["policy"] = "authorized_addresses"
+        self.cfg[0]["cfg"]["filtering"]["next_hop"]["authorized_addresses_list"] = ["192.168.0.1"]
+        self._contains_err()
+
     def test_multiple_ip_addresses(self):
         """{}: clients with multiple IP addresses"""
         clients_config = [
@@ -137,11 +173,22 @@ class TestConfigParserClients(TestConfigParserBase):
             "        ipv6_pref_len:",
             "          min: 3",
             "          max: 4",
+            "        # test pre v0.6.0 format for next_hop",
             "        next_hop_policy: same-as",
             "        max_prefix:",
             "          peering_db: False",
             "          limit_ipv4: 10",
-            "          limit_ipv6: 20"
+            "          limit_ipv6: 20",
+            "  - asn: 333",
+            "    ip: 192.0.2.31",
+            "    cfg:",
+            "      filtering:",
+            "        next_hop:",
+            "          policy: 'authorized_addresses'",
+            "          authorized_addresses_list:",
+            "            - '192.0.2.31'",
+            "            - '192.0.2.32'",
+            "            - '2001:db8:0:0::31'"
         ]))
         self.cfg.parse()
         self._contains_err()
@@ -160,7 +207,7 @@ class TestConfigParserClients(TestConfigParserBase):
         self.assertEqual(client["cfg"]["filtering"]["ipv4_pref_len"]["max"], 24)
         self.assertEqual(client["cfg"]["filtering"]["ipv6_pref_len"]["min"], 12)
         self.assertEqual(client["cfg"]["filtering"]["ipv6_pref_len"]["max"], 48)
-        self.assertEqual(client["cfg"]["filtering"]["next_hop_policy"], "strict")
+        self.assertEqual(client["cfg"]["filtering"]["next_hop"]["policy"], "strict")
         self.assertEqual(client["cfg"]["filtering"]["max_prefix"]["limit_ipv4"], None)
         self.assertEqual(client["cfg"]["filtering"]["max_prefix"]["limit_ipv6"], None)
         self.assertEqual(client["cfg"]["filtering"]["max_prefix"]["peering_db"], True)
@@ -179,10 +226,16 @@ class TestConfigParserClients(TestConfigParserBase):
         self.assertEqual(client["cfg"]["filtering"]["ipv4_pref_len"]["max"], 2)
         self.assertEqual(client["cfg"]["filtering"]["ipv6_pref_len"]["min"], 3)
         self.assertEqual(client["cfg"]["filtering"]["ipv6_pref_len"]["max"], 4)
-        self.assertEqual(client["cfg"]["filtering"]["next_hop_policy"], "same-as")
+        self.assertEqual(client["cfg"]["filtering"]["next_hop"]["policy"], "same-as")
         self.assertEqual(client["cfg"]["filtering"]["max_prefix"]["limit_ipv4"], 10)
         self.assertEqual(client["cfg"]["filtering"]["max_prefix"]["limit_ipv6"], 20)
         self.assertEqual(client["cfg"]["filtering"]["max_prefix"]["peering_db"], False)
+
+        client = self.cfg[2]
+        self.assertEqual(client["cfg"]["filtering"]["next_hop"]["policy"], "authorized_addresses")
+        self.assertEqual(client["cfg"]["filtering"]["next_hop"]["authorized_addresses_list"], [
+            "192.0.2.31", "192.0.2.32", "2001:db8:0:0::31"
+        ])
 
     def test_blackhole_filtering_propagation(self):
         """{}: inherit from general cfg - blackhole filtering"""